fix: amazon-ami-management plugin (#57)
Recent versions of Packer have validation to ensure that the version of the plugin reported via the `describe` command match the version in the filename. Our fork of the `amazon-ami-management` plugin had its version set to `2.0.0-dev`, but the filename just had `2.0.0`.

The reason we had forked was to add the ability to filter by tags. This functionality is available in the upstream, which also has the fix for the version number, so we can switch back.