ci: switch to OIDC Federation for Azure (#60)
Updating the packer build for Azure to use OIDC Federation instead of a static credential.

I've updated the CI build to have valid credentials for a service principal. This new SP has no access to anything - I've just added it to allow the packer validate step to work. I've also just hard-coded the client IDs into the actions on the grounds that they aren't credentials and don't need to be in secrets.